feat/#98: 유저 프로필 파싱 오류 수정 대기

+++ app/src/main/java/com/paw/key/presentation/ui/mypage/ArchivedCourseListScreen.kt
@@ -59,7 +59,7 @@ fun ArchivedCourseListScreen(
                 items = state.courseList
             ) { _, item ->
                 CourseCard(
-                    postId = item.postId,
+                    postId = item.postId.toInt(),
                     title = item.title,
                     createdAt = item.createdAt,
                     representativeImageUrl = item.representativeImageUrl,
